changed output directory for new sequences
Previously, all sequences from the same gene were appended to one file.
This is a misguided approach, because if the script gets rerun after something changes,
I will end up with two sequences from the same species in each file.
Now, separate files are made for each species and they are overwritten if the script is rerun