changed the dependancies for the blobology and all rules
In theory, Analyses/Trees is dependent on Analyses/Alignments/Phylip/AllConcatinatedProt.phy
so I replaced the latter with the former in the all rule

Blobology is now dependent on the assemblies, which are in turn dependent on the reads, so it
shouldn't be necessary for blobology to depend directly on the reads.

The tricky part here is that "%-scaffolds.fa Reads/%_1.fastq" rule should apply whenever
the reads exist, but the JGI download rules should substitute in other cases.

I think it should be OK for the annotation step to be dependent on the Assembly step, because
the AA translations will always be downloaded from JGI right after the assemblies, so they
should always be more recent